API::ADD: new API to pilot the LoRa module (#18)
* LORA_API::ADD: creation of lora API to handle reception and transmission

* Z1::CHANGE: use of the new API

* Z1::ADD: handle cypher

* ENCYPHER::FIX: payload size set to fix value

* API::CHANGE: transmission routine

* API::FIX: clen and mlen allocation, optimization in the dialog between the API and the Python

+++ Lora/src/ascon/crypto_aead.h
@@ -1,11 +1,11 @@
-int crypto_aead_encrypt(unsigned char *c, unsigned long long *clen,
-                        const unsigned char *m, unsigned long long mlen,
+int crypto_aead_encrypt(unsigned char *c, uint8_t *clen,
+                        const unsigned char *m, uint8_t mlen,
                         const unsigned char *ad, unsigned long long adlen,
                         const unsigned char *nsec, const unsigned char *npub,
                         const unsigned char *k);
 
-int crypto_aead_decrypt(unsigned char *m, unsigned long long *mlen,
+int crypto_aead_decrypt(unsigned char *m, uint8_t *mlen,
                         unsigned char *nsec, const unsigned char *c,
-                        unsigned long long clen, const unsigned char *ad,
+                        uint8_t clen, const unsigned char *ad,
                         unsigned long long adlen, const unsigned char *npub,
                         const unsigned char *k);
